services/chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pb",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "openrouter",
    "PyYAML>=6.0.0"
]

[tool.uv.sources]
pb = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
